Moving the settings to main paperless application
