+++ src/paperless/views.py
@@ -18,7 +18,9 @@ from rest_framework.viewsets import ModelViewSet
 from documents.permissions import PaperlessObjectPermissions
 from paperless.filters import GroupFilterSet
 from paperless.filters import UserFilterSet
+from paperless.models import OcrSettings
 from paperless.serialisers import GroupSerializer
+from paperless.serialisers import OcrSettingsSerializer
 from paperless.serialisers import ProfileSerializer
 from paperless.serialisers import UserSerializer
 
@@ -160,3 +162,12 @@ class GenerateAuthTokenView(GenericAPIView):
         return Response(
             token.key,
         )
+
+
+class OcrSettingsViewSet(ModelViewSet):
+    model = OcrSettings
+
+    queryset = OcrSettings.objects
+
+    serializer_class = OcrSettingsSerializer
+    permission_classes = (IsAuthenticated,)
