--- src/paperless_tesseract/serialisers.py
@@ -1,9 +1,0 @@
-from rest_framework import serializers
-
-from paperless_tesseract.models import OcrSettings
-
-
-class OcrSettingsSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = OcrSettings
-        fields = ["all"]
